Split up grid item components and add stories
