+++ client/src/components/grid/index.test.js
@@ -3,8 +3,8 @@ import React from 'react';
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
-import Grid from './grid';
-import GridItem from './grid-item';
+import Grid from '.';
+import GridItem from '../grid-item';
 import { gridItemData } from '../../test-helpers/_mock-data';
 
 configure({ adapter: new Adapter() });
