+++ client/src/components/index.js
@@ -1,5 +1,5 @@
 export {default as Header} from './header';
 export {default as Filter} from './filter/filter';
 export {default as CenterContent} from './center-content';
-export {default as Grid} from './grid/grid';
-export {default as GridItem} from './grid/grid-item';
+export {default as Grid} from './grid';
+export {default as GridItem} from './grid-item';
